+++ src/classes/ressources/PizzaSize.php
@@ -5,7 +5,7 @@ namespace classes\ressources;
 class PizzaSize
 {
     public $size, $price;
-    public function __construct(string $size, double $price)
+    public function __construct(string $size, float $price)
     {
         $this->size = $size;
         $this->price = $price;
